# backend/contracts/openapi.yaml
# PLATFORM-14 · contract (OpenAPI 3.1) · Код + тест · 2026-09-02
# GENERATED. Do not edit by hand.
#
# This document is produced by the running application and written through the one
# hand-owned normaliser, backend/src/test/java/mn/netgroup/cb/productcatalog/contract/
# OpenApiNormalizer. OpenApiContractDriftTest regenerates it twice, under two different
# default time zones and locales, and fails ./mvnw verify on any byte difference against
# this file. A difference is fixed in the code, never by editing this file to match.
#
# It is structurally identical - paths, operations, parameters, schemas, responses and
# error codes - to the contract approved at the design gate for 001-product-family-catalog.
components:
  parameters:
    FamilyId:
      description: The opaque identifier the service assigned to the family.
      examples:
        assigned:
          value: 0192f3a1-6c7d-7c3e-9f21-4b7c1d2e3f40
      in: path
      name: id
      required: true
      schema:
        format: uuid
        type: string
  responses:
    FamilyNotFound:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: "No persisted family holds that opaque identifier (FAMILY_NOT_FOUND). A path segment that is not a well-formed identifier is one no family holds and is answered here, with the same body as a well-formed identifier that is simply absent: every member that could discriminate the two is identical, and only instance differs, which is the caller's own request URI reflected back."
    InternalError:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: "An unhandled failure occurred (INTERNAL_ERROR). The body carries a correlation identifier and no other internal detail - no exception message, no class name, no stack frame."
  schemas:
    CreateProductFamilyRequest:
      description: "A request to register a product family. The status, the opaque identifier and both instants are the service's to assign and are not accepted here."
      properties:
        familyCode:
          maxLength: 20
          minLength: 3
          pattern: "^[A-Z0-9]{3,20}$"
          type: string
        name:
          maxLength: 120
          minLength: 1
          type: string
      required:
      - familyCode
      - name
      type: object
    ErrorCode:
      description: "The closed catalog of machine-readable error codes. Adding, removing or re-typing a member is a change to this contract and to the committed catalog snapshot."
      enum:
      - FAMILY_CODE_INVALID
      - FAMILY_NAME_INVALID
      - FAMILY_CODE_DUPLICATE
      - FAMILY_NOT_FOUND
      - LIMIT_ABOVE_MAXIMUM
      - CURSOR_INVALID
      - STATUS_FILTER_INVALID
      - INTERNAL_ERROR
      type: string
    HealthStatus:
      description: "The readiness of the service, as Spring Boot Actuator reports it."
      properties:
        status:
          type: string
      required:
      - status
      type: object
    Problem:
      description: "An RFC 9457 problem document. Every error response declared on the operations in this document has this shape, is built in one place, and carries a code from one catalog, so a client may branch on code alone. Clients branch on code and never on title or detail prose. Unknown members are ignored by RFC 9457, so a later code or parameter is an additive change. Scope, stated because the difference is not visible from the schema: the framework-produced failures named in info.description - 400 on a body that is not readable JSON, 405, 406, 415 - are problem documents of this shape with no code member, and they are not declared as responses on any operation here. A generated client must not assume that every problem+json body it can receive satisfies this schema until the design's OI-006 is answered."
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        correlationId:
          description: "Present on INTERNAL_ERROR only. The identifier that retrieves the log event for this failure, and the only internal detail a 500 exposes."
          format: uuid
          type: string
        detail:
          description: Human-readable prose. Not stable; never branch on it.
          type: string
        instance:
          format: uri-reference
          type: string
        max:
          description: Present on LIMIT_ABOVE_MAXIMUM only. The highest limit the contract admits.
          format: int32
          type: integer
        min:
          description: Present on LIMIT_ABOVE_MAXIMUM only. The lowest limit the contract admits.
          format: int32
          type: integer
        status:
          format: int32
          type: integer
        title:
          description: Human-readable prose. Not stable; never branch on it.
          type: string
        type:
          default: about:blank
          format: uri
          type: string
        violation:
          description: "Present on LIMIT_ABOVE_MAXIMUM only. Which bound the limit broke, so a client library can repair the request rather than retry it unchanged."
          enum:
          - ABOVE_MAX
          - BELOW_MIN
          - NOT_AN_INTEGER
          type: string
      required:
      - code
      - status
      type: object
    ProductFamily:
      description: A named grouping under which product definitions are later registered.
      properties:
        createdAt:
          description: "The instant the service first persisted the family, in UTC with a Z designator."
          format: date-time
          type: string
        familyCode:
          description: "The human-facing handle, supplied at creation, unique across the service, immutable, and stored exactly as supplied. Nothing may parse meaning out of it."
          maxLength: 20
          minLength: 3
          pattern: "^[A-Z0-9]{3,20}$"
          type: string
        id:
          description: "The opaque identifier the service assigned. Clients never supply or choose one, and nothing may be parsed out of it - it is not an ordering."
          format: uuid
          type: string
        name:
          maxLength: 120
          minLength: 1
          type: string
        status:
          $ref: "#/components/schemas/ProductFamilyStatus"
        updatedAt:
          description: "The instant the service last changed the family, in UTC with a Z designator."
          format: date-time
          type: string
      required:
      - createdAt
      - familyCode
      - id
      - name
      - status
      - updatedAt
      type: object
    ProductFamilyPage:
      description: "One keyset page. The response is flat - there is no data envelope, no _links member and no total count."
      properties:
        items:
          items:
            $ref: "#/components/schemas/ProductFamily"
          type: array
        nextCursor:
          description: "The cursor addressing the page after this one, or null when this is the last page. Null means end; a non-null value always fetches a further page."
          maxLength: 512
          type:
          - string
          - "null"
      required:
      - items
      - nextCursor
      type: object
    ProductFamilyStatus:
      description: The two states of a product family. ACTIVE is initial; RETIRED is terminal.
      enum:
      - ACTIVE
      - RETIRED
      type: string
info:
  description: |-
    The core-banking product family catalog. A product family carries an opaque identifier assigned by the service, a client-supplied immutable family code, a name, and a lifecycle status of ACTIVE or RETIRED. Families are created, read by identifier, listed a keyset page at a time, and retired. Retirement is terminal and idempotent; no operation returns a family to ACTIVE and no operation changes a persisted family code. This service declares no authentication: it returns no 401 and no 403 (spec 1, OI-005).

    Named gap, open at this gate: the protocol-level failures the framework produces before any operation below is entered - 400 on a body that is not readable JSON, 405, 406, 415 - are RFC 9457 problem documents with the right status, but the approved error-code catalog has no member for them, so they carry no code member. They are therefore not declared as responses on the operations below. See the design's OI-006 and decision D-08.
  title: Product family catalog
  version: 0.1.0
openapi: 3.1.0
paths:
  /actuator/health:
    get:
      description: "Spring Boot Actuator health, with the datasource indicator on. Read by the container orchestrator's probe and by the deploy stage. It returns no family data."
      operationId: health
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
          description: The service is able to serve requests.
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
          description: A dependency the service needs is not available.
      summary: Report whether the service is able to serve requests
      tags:
      - Operations
      x-requirements:
      - FR-014
  /v1/product-families:
    get:
      description: "Returns one keyset page ordered by familyCode ascending, with the opaque identifier appended as the final tiebreak only. The identifier is not part of the declared sort vocabulary, and the relative order of two families sharing a family code is an explicit non-promise of this contract - nothing may depend on it. There is no offset, page or pageNumber parameter and no total count. nextCursor is null only on the last page, and a non-null nextCursor always fetches a further page. A cursor is issued by the service, is integrity-sealed, and carries the sort specification it was issued for; clients pass one back unmodified and never construct or edit one. This page is not a snapshot: a family created after the first page may appear on a later one. With no status parameter, families of every status are returned (spec OI-003)."
      operationId: listProductFamilies
      parameters:
      - description: Return only families whose status equals this value.
        in: query
        name: status
        required: false
        schema:
          $ref: "#/components/schemas/ProductFamilyStatus"
      - description: "The maximum number of families in the page. It must be an integer between 1 and 100. Any value the declared bounds do not admit - above the maximum, below the minimum, absent-but-present as an empty value, repeated, not an integer, or beyond the range of a 32-bit integer - is rejected with 400 LIMIT_ABOVE_MAXIMUM carrying a violation member that says which. It is never silently clamped to the maximum and never silently defaulted."
        in: query
        name: limit
        required: false
        schema:
          default: 20
          format: int32
          maximum: 100
          minimum: 1
          type: integer
      - description: "The nextCursor of the previous page, passed back unmodified. It is integrity-sealed and carries the key identifier it was sealed under, so the sealing key can be rotated without invalidating cursors already issued. It is sealed, not confidential - a client can read the sort specification and the last row's values out of it, and cannot forge one - and forging one would grant no read the caller does not already have."
        in: query
        name: cursor
        required: false
        schema:
          maxLength: 512
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductFamilyPage"
          description: One page of product families.
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The limit is not an integer between 1 and 100 (LIMIT_ABOVE_MAXIMUM, with a violation member of ABOVE_MAX, BELOW_MIN or NOT_AN_INTEGER and the min and max members that bound it), the cursor failed its integrity check or was issued for a different sort specification (CURSOR_INVALID), or the status filter is neither ACTIVE nor RETIRED (STATUS_FILTER_INVALID)."
        "500":
          $ref: "#/components/responses/InternalError"
      summary: "List product families, one keyset page at a time"
      tags:
      - Product families
      x-requirements:
      - FR-006
      - FR-007
      - FR-008
      - FR-009
      - FR-020
      - FR-021
      - FR-022
      - FR-023
    post:
      description: Persists a new product family with status ACTIVE and an opaque identifier the service assigns. The family code is stored exactly as supplied and can never be changed afterwards.
      operationId: createProductFamily
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductFamilyRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductFamily"
          description: The family was persisted.
          headers:
            Location:
              description: "The created family, addressed by its opaque identifier."
              required: true
              schema:
                format: uri-reference
                type: string
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The family code is not 3 to 20 characters of A-Z and 0-9 (FAMILY_CODE_INVALID), or the name is not 1 to 120 characters (FAMILY_NAME_INVALID)."
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: A persisted family already holds that family code (FAMILY_CODE_DUPLICATE).
        "500":
          $ref: "#/components/responses/InternalError"
      summary: Register a product family
      tags:
      - Product families
      x-requirements:
      - FR-001
      - FR-002
      - FR-003
      - FR-004
      - FR-016
      - FR-017
      - FR-018
      - FR-023
  /v1/product-families/{id}:
    get:
      description: Returns the family addressed by the opaque identifier. The family code is not an identifier in a URL and cannot address a family here.
      operationId: getProductFamily
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductFamily"
          description: The family.
        "404":
          $ref: "#/components/responses/FamilyNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
      summary: Read one product family by its opaque identifier
      tags:
      - Product families
      x-requirements:
      - FR-005
      - FR-019
      - FR-023
    parameters:
    - $ref: "#/components/parameters/FamilyId"
  /v1/product-families/{id}/retire:
    parameters:
    - $ref: "#/components/parameters/FamilyId"
    post:
      description: "Transitions an ACTIVE family to RETIRED. RETIRED is terminal: a retire request against an already RETIRED family is idempotent and returns that family unchanged, including its updatedAt. There is no request body and no operation that returns a family to ACTIVE."
      operationId: retireProductFamily
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductFamily"
          description: "The family, now RETIRED."
        "404":
          $ref: "#/components/responses/FamilyNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
      summary: Retire a product family
      tags:
      - Product families
      x-requirements:
      - FR-010
      - FR-011
      - FR-013
      - FR-019
      - FR-023
tags:
- description: "Registration, retrieval, listing and retirement of product families."
  name: Product families
- description: Deployment readiness.
  name: Operations